After 3 days banging my head against this error:
[ERROR] Failed to execute goal org.apache.maven.plugins:maven-shade-plugin:1.7.1
:shade (default) on project logwrapper: Error creating shaded jar: error in open
ing zip file C:\Users\bruno.gonzalez\.m2\repository\org\wixtoolset\maven\wix-too
lset\3.9.2\wix-toolset-3.9.2-WixUIExtension.wixext -> [Help 1]

We are swtiching to use the Inno compiler. Basically this erros is that the zip file is not there, I have debug the shade plugin in Maven and it is just looking for a dll on that location. There is a file with the name with that name
but not zip extension that can be unzipped but does not look like it has anything inside.

It maybe a repository issue or that I don't know how to use the shade pulgin correctly with the wix toolset Maven plugin dependencies.